refactor(asset): change asset mixin to asset serializer
- move asset model and serializers to "common" app
