+++ apps/project/models.py
@@ -10,53 +10,12 @@ from django.db.models.functions import Concat, Lower
 from django.utils.translation import gettext_lazy
 from django_choices_field import IntegerChoicesField
 
-from apps.common.models import ArchivableResource, FirebasePushStatusEnum, FirebaseResource, UserResource
+from apps.common.models import ArchivableResource, CommonAsset, FirebasePushStatusEnum, FirebaseResource, UserResource
 from apps.contributor.models import ContributorTeam
 from utils.fields import validate_percentage
 
 if typing.TYPE_CHECKING:
     from apps.tutorial.models import Tutorial
-
-
-class ProjectAssetMimetypeEnum(models.IntegerChoices):
-    GEOJSON = 100, "application/geo+json"
-
-    IMAGE_JPEG = 201, "image/jpeg"
-    IMAGE_PNG = 202, "image/png"
-    IMAGE_GIF = 203, "image/gif"
-
-    @classmethod
-    def get_display(cls, value: typing.Self | int) -> str:
-        if value in cls:
-            return str(cls(value).label)
-        return "Unknown"
-
-    @classmethod
-    def is_valid_mimetype(cls, mimetype: str) -> bool:
-        """
-        Check if the given mimetype is valid for project assets.
-        """
-        return mimetype in [choice.label for choice in cls]
-
-    @classmethod
-    def get_mimetype_by_label(cls, label: str) -> typing.Self | None:
-        for choice in cls:
-            if choice.label == label:
-                return choice
-        return None
-
-
-# FIXME(tnagorra): Finalize the enum labels
-class ProjectAssetTypeEnum(models.IntegerChoices):
-    INPUT = 100, "Input"
-    OUTPUT = 200, "Output"
-    STATS = 300, "Stats"
-
-    @classmethod
-    def get_display(cls, value: typing.Self | int) -> str:
-        if value in cls:
-            return str(cls(value).label)
-        return "Unknown"
 
 
 class ProjectTypeEnum(models.IntegerChoices):
@@ -421,43 +380,17 @@ class Project(UserResource, FirebaseResource):  # type: ignore[reportIncompatibl
         #     self.requiredResults += group.requiredCount * group.numberOfTasks
 
 
-class ProjectAsset(UserResource):
-    Type = ProjectAssetTypeEnum
-    Mimetype = ProjectAssetMimetypeEnum
-    MAX_FILE_SIZE: int = 10 * 1024 * 1024  # MB
-
-    type = IntegerChoicesField(
-        choices_enum=ProjectAssetTypeEnum,
-    )
-
-    mimetype = IntegerChoicesField(
-        choices_enum=ProjectAssetMimetypeEnum,
-    )
-
-    file = models.FileField(
-        upload_to=UploadHelper.project_asset,
-        help_text=gettext_lazy("The file associated with the asset"),
-    )
-
-    file_size = models.PositiveIntegerField(
-        help_text=gettext_lazy("The size of the file in bytes"),
-    )
-
+class ProjectAsset(UserResource, CommonAsset):  # type: ignore[reportIncompatibleVariableOverride]
     project: Project = models.ForeignKey(  # type: ignore[reportAssignmentType]
         Project,
         on_delete=models.CASCADE,
         related_name="+",
     )
 
-    marked_as_deleted = models.BooleanField(
-        default=False,
-        help_text=gettext_lazy("If this flag is enabled, this project asset will be deleted in the future"),
+    file = models.FileField(
+        upload_to=UploadHelper.project_asset,
+        help_text=gettext_lazy("The file associated with the asset"),
     )
-
-    @classmethod
-    def usable_objects(cls):
-        """Returns objects that are mot marked for deletion"""
-        return cls.objects.filter(marked_as_deleted=False)
 
     # Type hints
     project_id: int
