+++ project_types/tile_map_service/base/project.py
@@ -13,11 +13,13 @@ from pyfirebase_mapswipe import extended_models as firebase_ext_models
 from pyfirebase_mapswipe import models as firebase_models
 from ulid import ULID
 
+from apps.common.models import (
+    AssetMimetypeEnum,
+    AssetTypeEnum,
+)
 from apps.project.models import (
     Project,
     ProjectAsset,
-    ProjectAssetMimetypeEnum,
-    ProjectAssetTypeEnum,
     ProjectTask,
     ProjectTaskGroup,
 )
@@ -47,8 +49,8 @@ class TileMapServiceProjectProperty(base_project.BaseProjectProperty):
             ProjectAsset.usable_objects()
             .filter(
                 id=self.aoi_geometry,
-                type=ProjectAssetTypeEnum.INPUT,
-                mimetype=ProjectAssetMimetypeEnum.GEOJSON,
+                type=AssetTypeEnum.INPUT,
+                mimetype=AssetMimetypeEnum.GEOJSON,
                 project_id=project_id,
             )
             .exists()
@@ -133,8 +135,8 @@ class TileMapServiceBaseProject[
             project=self.project,
             file=file,
             file_size=file.size,
-            type=ProjectAssetTypeEnum.OUTPUT,
-            mimetype=ProjectAssetMimetypeEnum.GEOJSON,
+            type=AssetTypeEnum.OUTPUT,
+            mimetype=AssetMimetypeEnum.GEOJSON,
             # FIXME(tnagorra): Maybe create a internal user like mapswipe-bot
             created_by=self.project.modified_by,
             modified_by=self.project.modified_by,
@@ -212,8 +214,8 @@ class TileMapServiceBaseProject[
 
         aoi_asset = ProjectAsset.usable_objects().get(
             id=self.project_type_specifics.aoi_geometry,
-            type=ProjectAssetTypeEnum.INPUT,
-            mimetype=ProjectAssetMimetypeEnum.GEOJSON,
+            type=AssetTypeEnum.INPUT,
+            mimetype=AssetMimetypeEnum.GEOJSON,
             project_id=self.project.pk,
         )
